diskimage: Use correct casting for comparing against (ULONG)-1
On x86_64, -1UL is 0xffffffffffffffff, not 0xffffffff


git-svn-id: https://svn.aros.org/svn/aros/trunk@52041 fb15a70f-31f2-0310-bbcc-cdcc74a49acc

+++ workbench/devs/diskimage/device/plugins.c
@@ -169,7 +169,7 @@ static ULONG InitPlugins (struct DiskImageBase *libBase, struct DiskImagePluginT
 			plugin->SegList = seglist;
 			plugin->RefCount = &plugin_table->RefCount;
 			if (!plugin->plugin_Init || Plugin_Init(plugin, plugin_data)) {
-				if (plugin_table->RefCount == -1UL) {
+				if (plugin_table->RefCount == (ULONG)-1) {
 					plugin->Flags |= PLUGIN_FLAG_BUILTIN;
 				} else {
 					plugin_table->RefCount++;
